requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alibabacloud-ocr-api20210707==3.1.2
alibabacloud-tea==0.4.0
alibabacloud_credentials==0.3.6
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.2
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.12
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
configparser==7.1.0
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
fonttools==4.55.3
frozenlist==1.5.0
idna==3.10
kiwisolver==1.4.8
marshmallow==3.26.1
matplotlib==3.10.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
pycparser==2.22
pyparsing==3.2.0
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
PySide6==6.8.1
PySide6_Addons==6.8.1
PySide6_Essentials==6.8.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
shiboken6==6.8.1
six==1.17.0
soupsieve==2.6
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
yarl==1.18.3
